test: Disable hostfw by default when running tests locally
The host firewall is disabled by default in the CI (unless you have
label ci/host-firewall on your PR). The same should be true when running
tests locally with the test VMs.

os.Getenv("HOST_FIREWALL") returns an empty string for undefined env.
variables, so we also need to match on that.

Signed-off-by: Paul Chaignon <paul@cilium.io>